+++ tomcat/src/main/java/nextstep/jwp/controller/base/Controller.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.handler.controller.base;
+package nextstep.jwp.controller.base;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Controller {
 
-    HttpResponse service(HttpRequest request) throws Exception;
+    HttpResponse handle(HttpRequest request) throws Exception;
 }
