+++ tomcat/src/main/java/nextstep/jwp/exception/UnsupportedMethodException.java
@@ -1,0 +1,8 @@
+package nextstep.jwp.exception;
+
+public class UnsupportedMethodException extends RuntimeException {
+
+    public UnsupportedMethodException() {
+        super("요청 메서드가 잘못 되었습니다.");
+    }
+}
