+++ tomcat/src/main/java/org/apache/coyote/http11/request/header/Params.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.header;
 
 import java.net.URLDecoder;
 import java.util.Arrays;
